refactor(MyHttpRequest): HttpRequest 확인에 대한 책임을 RequestHandler로부터 분리
