+++ web-application-server-gradle/src/main/java/webserver/RequestHandler.java
@@ -10,9 +10,7 @@ import webserver.http.Response;
 import java.io.*;
 import java.net.Socket;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class RequestHandler extends Thread {
     private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
@@ -30,7 +28,7 @@ public class RequestHandler extends Thread {
 
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             // BufferedReader 값 읽기
-            MyHttpRequest myHttpRequest = httpRequestFromInputStream(in);
+            MyHttpRequest myHttpRequest = new MyHttpRequest(in);
             log.debug("request  ===========>" + myHttpRequest.toString());
 
             // 컨트롤러 맵핑 및 반환
@@ -47,69 +45,6 @@ public class RequestHandler extends Thread {
         } catch (IOException e) {
             log.error(e.getMessage());
         }
-    }
-
-    private static MyHttpRequest httpRequestFromInputStream(InputStream in) throws IOException {
-        // BufferedReader 값 읽기
-        InputStreamReader reader = new InputStreamReader(in);
-        BufferedReader br = new BufferedReader(reader);
-
-        List<String> lines = readInputStream(br);
-
-        // RequestLine 가져오기
-        String[] tokens = lines.get(0).split(" ");
-        lines.remove(0);
-        String httpMethod = tokens[0];
-        String url = tokens[1];
-
-        // request path & parameter 읽기
-        String requestPath = url;
-        Map<String, String> parameters = null;
-        int index = url.indexOf("?");
-        if (index != -1) {
-            requestPath = url.substring(0, index);
-            String params = url.substring(index + 1);
-
-            // parameter map으로 분리
-            parameters = HttpRequestUtils.parseQueryString(params);
-        }
-
-        // header 읽기
-        Map<String, String> headers = new HashMap<String, String>();
-        for (String l : lines) {
-            if ("".equals(l)) {
-                break;
-            }
-            String[] headerTokens = l.split(": ");
-            if (headerTokens.length == 2) {
-                headers.put(headerTokens[0], headerTokens[1]);
-            }
-        }
-        log.debug("contentLength:" + headers.get("Content-Length"));
-
-        // request body 읽기
-        String requestBody = null;
-        if (headers.get("Content-Length") != null) {
-            requestBody = IOUtils.readData(br, Integer.parseInt(headers.get("Content-Length")));
-            if (!"".equals(requestBody)) {
-                log.debug("requestBody: " + requestBody);
-                if ("POST".equals(httpMethod)) {
-                    parameters = HttpRequestUtils.parseQueryString(requestBody);
-                }
-            }
-        }
-
-        return new MyHttpRequest(httpMethod, requestPath, parameters, headers, requestBody);
-    }
-
-    private static List<String> readInputStream(BufferedReader br) throws IOException {
-        List<String> lines = new ArrayList<>();
-        String line;
-        while ((line = br.readLine()) != null && !"".equals(line)) {
-            lines.add(line);
-            log.debug("BufferedReader: " + line);
-        }
-        return lines;
     }
 
     private void responseHeader(DataOutputStream dos, List<String> headers) {
